RealEstatePriceForecastBackend/requirements.txt
asgiref==3.8.1
python-dotenv
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
django-cors-headers==4.5.0
djangorestframework==3.15.2
fonttools==4.54.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psycopg2-binary==2.9.6
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.2
xgboost==2.1.1
